+++ tests/unit/actions/test_snap_installer.py
@@ -17,6 +17,7 @@
 
 import json
 import pathlib
+import subprocess
 import textwrap
 from unittest import mock
 
@@ -74,13 +75,13 @@ def config_fixture(request, tmp_path, mocker):
     )
 
 
-@pytest.fixture(params=["2"])
-def mock_get_host_snap_revision(request, mocker):
+@pytest.fixture(params=[{"revision": "2", "id": "3"}])
+def mock_get_host_snap_info(request, mocker):
     """Mocks the get_host_snap_revision() function
 
     :param request: The revision of the host's snap. Default revision = 2.
     """
-    mockee = "craft_providers.actions.snap_installer._get_host_snap_revision"
+    mockee = "craft_providers.actions.snap_installer.get_host_snap_info"
     if isinstance(request.param, Exception):
         mocker.patch(mockee, side_effect=request.param)
     else:
@@ -113,16 +114,23 @@ def mock_get_target_snap_revision_from_snapd(request, mocker):
 
 def test_inject_from_host_classic(
     config_fixture,
-    mock_get_host_snap_revision,
+    mock_get_host_snap_info,
     mock_requests,
     fake_executor,
     fake_process,
     logs,
     tmp_path,
 ):
-
+    # register 'snap known' calls
+    for _ in range(3):
+        fake_process.register_subprocess(["snap", "known", fake_process.any()])
     fake_process.register_subprocess(
-        ["fake-executor", "rm", "-f", "/tmp/test-name.snap"]
+        [
+            "fake-executor",
+            "snap",
+            "ack",
+            "/tmp/test-name.assert",
+        ]
     )
     fake_process.register_subprocess(
         [
@@ -131,7 +139,6 @@ def test_inject_from_host_classic(
             "install",
             "/tmp/test-name.snap",
             "--classic",
-            "--dangerous",
         ]
     )
 
@@ -139,14 +146,11 @@ def test_inject_from_host_classic(
         executor=fake_executor, snap_name="test-name", classic=True
     )
 
-    assert mock_requests.mock_calls == [
-        mock.call.get("http+unix://%2Frun%2Fsnapd.socket/v2/snaps/test-name/file"),
-        mock.call.get().raise_for_status(),
-        mock.call.get().iter_content(65536),
-        mock.call.get().iter_content().__iter__(),
-    ]
+    mock_requests.get.assert_called_with(
+        "http+unix://%2Frun%2Fsnapd.socket/v2/snaps/test-name/file"
+    )
 
-    assert len(fake_process.calls) == 2
+    assert len(fake_process.calls) == 5
     assert Exact("Installing snap 'test-name' from host (classic=True)") in logs.debug
     assert "Revisions found: host='2', target='1'" in logs.debug
 
@@ -166,16 +170,69 @@ def test_inject_from_host_classic(
 
 def test_inject_from_host_strict(
     config_fixture,
-    mock_get_host_snap_revision,
+    mock_get_host_snap_info,
     mock_requests,
     fake_executor,
     fake_process,
     logs,
     tmp_path,
 ):
+    # register 'snap known' calls
+    for _ in range(3):
+        fake_process.register_subprocess(["snap", "known", fake_process.any()])
     fake_process.register_subprocess(
-        ["fake-executor", "rm", "-f", "/tmp/test-name.snap"]
+        [
+            "fake-executor",
+            "snap",
+            "ack",
+            "/tmp/test-name.assert",
+        ]
     )
+    fake_process.register_subprocess(
+        [
+            "fake-executor",
+            "snap",
+            "install",
+            "/tmp/test-name.snap",
+        ]
+    )
+
+    snap_installer.inject_from_host(
+        executor=fake_executor, snap_name="test-name", classic=False
+    )
+
+    mock_requests.get.assert_called_with(
+        "http+unix://%2Frun%2Fsnapd.socket/v2/snaps/test-name/file"
+    )
+
+    assert len(fake_process.calls) == 5
+    assert Exact("Installing snap 'test-name' from host (classic=False)") in logs.debug
+    assert "Revisions found: host='2', target='1'" in logs.debug
+
+    # check saved config
+    (saved_config_record,) = [
+        x
+        for x in fake_executor.records_of_push_file_io
+        if "craft-instance.conf" in x["destination"]
+    ]
+    config = InstanceConfiguration(**yaml.safe_load(saved_config_record["content"]))
+    assert config.snaps is not None
+    assert config.snaps["test-name"] == {  # pylint: disable=unsubscriptable-object
+        "revision": "2",
+        "source": snap_installer.SNAP_SRC_HOST,
+    }
+
+
+@pytest.mark.parametrize("mock_get_host_snap_info", [{"revision": "x3"}], indirect=True)
+def test_inject_from_host_dangerous(
+    config_fixture,
+    mock_get_host_snap_info,
+    mock_requests,
+    fake_executor,
+    fake_process,
+    logs,
+    tmp_path,
+):
     fake_process.register_subprocess(
         [
             "fake-executor",
@@ -190,14 +247,89 @@ def test_inject_from_host_strict(
         executor=fake_executor, snap_name="test-name", classic=False
     )
 
-    assert mock_requests.mock_calls == [
-        mock.call.get("http+unix://%2Frun%2Fsnapd.socket/v2/snaps/test-name/file"),
-        mock.call.get().raise_for_status(),
-        mock.call.get().iter_content(65536),
-        mock.call.get().iter_content().__iter__(),
-    ]
+    mock_requests.get.assert_called_with(
+        "http+unix://%2Frun%2Fsnapd.socket/v2/snaps/test-name/file"
+    )
 
-    assert len(fake_process.calls) == 2
+    assert len(fake_process.calls) == 1
+    assert Exact("Installing snap 'test-name' from host (classic=False)") in logs.debug
+    assert "Revisions found: host='x3', target='1'" in logs.debug
+
+    # check saved config
+    (saved_config_record,) = [
+        x
+        for x in fake_executor.records_of_push_file_io
+        if "craft-instance.conf" in x["destination"]
+    ]
+    config = InstanceConfiguration(**yaml.safe_load(saved_config_record["content"]))
+    assert config.snaps is not None
+    assert config.snaps["test-name"] == {  # pylint: disable=unsubscriptable-object
+        "revision": "x3",
+        "source": snap_installer.SNAP_SRC_HOST,
+    }
+
+
+def test_inject_from_host_not_dangerous(
+    config_fixture,
+    mock_get_host_snap_info,
+    mock_requests,
+    fake_executor,
+    fake_process,
+    logs,
+    tmp_path,
+):
+    fake_process.register_subprocess(
+        [
+            "snap",
+            "known",
+            "account-key",
+            "public-key-sha3-384=BWDEoaqyr25nF5SNCvEv2v"
+            "7QnM9QsfCc0PBMYD_i2NGSQ32EF2d4D0hqUel3m8ul",
+        ]
+    )
+    fake_process.register_subprocess(
+        [
+            "snap",
+            "known",
+            "snap-declaration",
+            "snap-name=test-name",
+        ]
+    )
+    fake_process.register_subprocess(
+        [
+            "snap",
+            "known",
+            "snap-revision",
+            "snap-revision=2",
+            "snap-id=3",
+        ]
+    )
+    fake_process.register_subprocess(
+        [
+            "fake-executor",
+            "snap",
+            "ack",
+            "/tmp/test-name.assert",
+        ]
+    )
+    fake_process.register_subprocess(
+        [
+            "fake-executor",
+            "snap",
+            "install",
+            "/tmp/test-name.snap",
+        ]
+    )
+
+    snap_installer.inject_from_host(
+        executor=fake_executor, snap_name="test-name", classic=False
+    )
+
+    mock_requests.get.assert_called_with(
+        "http+unix://%2Frun%2Fsnapd.socket/v2/snaps/test-name/file"
+    )
+
+    assert len(fake_process.calls) == 5
     assert Exact("Installing snap 'test-name' from host (classic=False)") in logs.debug
     assert "Revisions found: host='2', target='1'" in logs.debug
 
@@ -216,10 +348,10 @@ def test_inject_from_host_strict(
 
 
 @pytest.mark.parametrize("config_fixture", ["10"], indirect=True)
-@pytest.mark.parametrize("mock_get_host_snap_revision", ["10"], indirect=True)
+@pytest.mark.parametrize("mock_get_host_snap_info", [{"revision": "10"}], indirect=True)
 def test_inject_from_host_matching_revision_no_op(
     config_fixture,
-    mock_get_host_snap_revision,
+    mock_get_host_snap_info,
     mock_requests,
     fake_executor,
     fake_process,
@@ -239,11 +371,11 @@ def test_inject_from_host_matching_revision_no_op(
 
 
 @pytest.mark.parametrize(
-    "mock_get_host_snap_revision",
+    "mock_get_host_snap_info",
     [snap_installer.SnapInstallationError("msg")],
     indirect=True,
 )
-def test_inject_from_host_no_snapd(mock_get_host_snap_revision, fake_executor):
+def test_inject_from_host_no_snapd(mock_get_host_snap_info, fake_executor):
     """The host does not have snapd at all."""
     with pytest.raises(snap_installer.SnapInstallationError):
         snap_installer.inject_from_host(
@@ -257,25 +389,21 @@ def test_inject_from_host_push_error(
     mock_executor = mock.Mock(spec=fake_executor, wraps=fake_executor)
     mock_executor.push_file.side_effect = ProviderError(brief="foo")
 
-    fake_process.register_subprocess(
-        ["fake-executor", "rm", "-f", "/tmp/test-name.snap"]
-    )
-
     with pytest.raises(snap_installer.SnapInstallationError) as exc_info:
         snap_installer.inject_from_host(
             executor=mock_executor, snap_name="test-name", classic=False
         )
 
     assert exc_info.value == snap_installer.SnapInstallationError(
-        brief="Failed to inject snap 'test-name'.",
-        details="Error copying snap into target environment.",
+        brief="failed to copy snap file for snap 'test-name'",
+        details="error copying snap file into target environment",
     )
     assert exc_info.value.__cause__ is not None
 
 
 def test_inject_from_host_snapd_connection_error_using_pack_fallback(
     config_fixture,
-    mock_get_host_snap_revision,
+    mock_get_host_snap_info,
     mock_requests,
     fake_executor,
     fake_process,
@@ -284,14 +412,22 @@ def test_inject_from_host_snapd_connection_error_using_pack_fallback(
     mock_requests.get.side_effect = requests.exceptions.ConnectionError()
 
     fake_process.register_subprocess(
-        ["fake-executor", "rm", "-f", "/tmp/test-name.snap"]
-    )
-    fake_process.register_subprocess(
         [
             "snap",
             "pack",
             "/snap/test-name/current/",
             f'--filename={pathlib.Path(tmpdir / "test-name.snap")}',
+        ]
+    )
+    # register 'snap known' calls
+    for _ in range(3):
+        fake_process.register_subprocess(["snap", "known", fake_process.any()])
+    fake_process.register_subprocess(
+        [
+            "fake-executor",
+            "snap",
+            "ack",
+            "/tmp/test-name.assert",
         ]
     )
     fake_process.register_subprocess(
@@ -300,7 +436,6 @@ def test_inject_from_host_snapd_connection_error_using_pack_fallback(
             "snap",
             "install",
             "/tmp/test-name.snap",
-            "--dangerous",
         ]
     )
 
@@ -311,12 +446,12 @@ def test_inject_from_host_snapd_connection_error_using_pack_fallback(
     assert mock_requests.mock_calls == [
         mock.call.get("http+unix://%2Frun%2Fsnapd.socket/v2/snaps/test-name/file"),
     ]
-    assert len(fake_process.calls) == 3
+    assert len(fake_process.calls) == 6
 
 
 def test_inject_from_host_snapd_http_error_using_pack_fallback(
     config_fixture,
-    mock_get_host_snap_revision,
+    mock_get_host_snap_info,
     mock_requests,
     fake_executor,
     fake_process,
@@ -326,14 +461,22 @@ def test_inject_from_host_snapd_http_error_using_pack_fallback(
         requests.exceptions.HTTPError()
     )
     fake_process.register_subprocess(
-        ["fake-executor", "rm", "-f", "/tmp/test-name.snap"]
-    )
-    fake_process.register_subprocess(
         [
             "snap",
             "pack",
             "/snap/test-name/current/",
             f'--filename={pathlib.Path(tmpdir / "test-name.snap")}',
+        ]
+    )
+    # register 'snap known' calls
+    for _ in range(3):
+        fake_process.register_subprocess(["snap", "known", fake_process.any()])
+    fake_process.register_subprocess(
+        [
+            "fake-executor",
+            "snap",
+            "ack",
+            "/tmp/test-name.assert",
         ]
     )
     fake_process.register_subprocess(
@@ -342,7 +485,6 @@ def test_inject_from_host_snapd_http_error_using_pack_fallback(
             "snap",
             "install",
             "/tmp/test-name.snap",
-            "--dangerous",
         ]
     )
 
@@ -355,15 +497,12 @@ def test_inject_from_host_snapd_http_error_using_pack_fallback(
         mock.call.get().raise_for_status(),
     ]
 
-    assert len(fake_process.calls) == 3
+    assert len(fake_process.calls) == 6
 
 
 def test_inject_from_host_install_failure(
     mock_requests, config_fixture, fake_executor, fake_process
 ):
-    fake_process.register_subprocess(
-        ["fake-executor", "rm", "-f", "/tmp/test-name.snap"]
-    )
     fake_process.register_subprocess(
         [
             "fake-executor",
@@ -381,13 +520,13 @@ def test_inject_from_host_install_failure(
         )
 
     assert exc_info.value == snap_installer.SnapInstallationError(
-        brief="Failed to inject snap 'test-name'.",
+        brief="failed to install snap 'test-name'",
         details=details_from_called_process_error(
             exc_info.value.__cause__  # type: ignore
         ),
     )
 
-    assert len(fake_process.calls) == 2
+    assert len(fake_process.calls) == 1
 
 
 @pytest.mark.parametrize(
@@ -567,7 +706,7 @@ def test_install_from_store_failure(
 # -- tests for the helping functions
 
 
-def test_get_host_snap_revision_ok(responses):
+def test_get_host_snap_info_ok(responses):
     """Revision retrieved ok."""
     snap_info = {"result": {"revision": "15"}}
     responses.add(
@@ -575,22 +714,97 @@ def test_get_host_snap_revision_ok(responses):
         "http+unix://%2Frun%2Fsnapd.socket/v2/snaps/test-snap",
         json=snap_info,
     )
-    result = snap_installer._get_host_snap_revision("test-snap")
+    result = snap_installer.get_host_snap_info("test-snap")["revision"]
     assert result == "15"
 
 
-def test_get_host_snap_revision_connection_error(responses):
+def test_get_host_snap_info_connection_error(responses):
     """Error when connecting to snapd.
 
     Note that nothing is added to responses, so ConnectionError will be raised when
     trying to connect, which is the effect we want.
     """
     with pytest.raises(snap_installer.SnapInstallationError) as exc_info:
-        snap_installer._get_host_snap_revision("test-snap")
+        snap_installer.get_host_snap_info("test-snap")
     assert exc_info.value == snap_installer.SnapInstallationError(
         brief="Unable to connect to snapd service."
     )
     assert exc_info.value.__cause__ is not None
+
+
+def test_get_assertion_connection_error(mocker):
+    """Raise SnapInstallation when 'snap known' call fails."""
+    mocker.patch(
+        "subprocess.run", side_effect=subprocess.CalledProcessError(100, ["error"])
+    )
+
+    with pytest.raises(snap_installer.SnapInstallationError) as exc_info:
+        snap_installer._get_assertion(["test1", "test2"])
+    assert exc_info.value == snap_installer.SnapInstallationError(
+        brief="failed to get assertions for snap",
+        details="* Command that failed: 'error'\n* Command exit code: 100",
+    )
+    assert exc_info.value.__cause__ is not None
+
+
+def test_add_assertions_from_host_error_on_push(
+    fake_executor, fake_process, mock_requests, mocker, tmpdir
+):
+    """Raise SnapInstallationError when assert file cannot be pushed."""
+    mock_executor = mock.Mock(spec=fake_executor, wraps=fake_executor)
+    mock_executor.push_file.side_effect = ProviderError(brief="foo")
+    mocker.patch(
+        "craft_providers.bases.instance_config.temp_paths.home_temporary_file",
+        return_value=pathlib.Path(tmpdir) / "temp-file",
+    )
+
+    # register 'snap known' calls
+    for _ in range(3):
+        fake_process.register_subprocess(["snap", "known", fake_process.any()])
+
+    with pytest.raises(snap_installer.SnapInstallationError) as exc_info:
+        snap_installer._add_assertions_from_host(
+            executor=mock_executor,
+            snap_name="test-name",
+        )
+
+    assert exc_info.value == snap_installer.SnapInstallationError(
+        brief="failed to copy assert file for snap 'test-name'",
+        details="error copying snap assert file into target environment",
+    )
+    assert exc_info.value.__cause__ is not None
+
+
+def test_add_assertions_from_host_error_on_ack(
+    fake_executor, fake_process, mock_requests, mocker, tmpdir
+):
+    """Raise SnapInstallationError when 'snap ack' fails."""
+    fake_process.register_subprocess(
+        ["fake-executor", "snap", "ack", "/tmp/test-name.assert"],
+        returncode=1,
+    )
+    mocker.patch(
+        "craft_providers.bases.instance_config.temp_paths.home_temporary_file",
+        return_value=pathlib.Path(tmpdir) / "temp-file",
+    )
+
+    # register 'snap known' calls
+    for _ in range(3):
+        fake_process.register_subprocess(["snap", "known", fake_process.any()])
+
+    with pytest.raises(snap_installer.SnapInstallationError) as exc_info:
+        snap_installer._add_assertions_from_host(
+            executor=fake_executor, snap_name="test-name"
+        )
+
+    assert exc_info.value == snap_installer.SnapInstallationError(
+        brief="failed to add assertions for snap 'test-name'",
+        details=details_from_called_process_error(
+            exc_info.value.__cause__  # type: ignore
+        ),
+    )
+
+    assert len(fake_process.calls) == 4
 
 
 def test_get_target_snap_revision_from_snapd_process_error(fake_process, fake_executor):