+++ tests/unit/bases/test_buildd.py
@@ -16,6 +16,7 @@
 #
 
 import subprocess
+from pathlib import Path
 from textwrap import dedent
 from unittest.mock import ANY, call, patch
 
@@ -46,6 +47,12 @@ def mock_load(mocker):
             compatibility_tag="buildd-base-v0"
         ),
     )
+
+
+# The variable name "fs" causes a pylint warning.  Provide a longer name.
+@pytest.fixture
+def fake_filesystem(fs):  # pylint: disable=invalid-name
+    yield fs
 
 
 @pytest.fixture()
@@ -90,6 +97,7 @@ def mock_inject_from_host(mocker):
         ),
     ],
 )
+@pytest.mark.parametrize("no_cdn", [False, True])
 @pytest.mark.parametrize(
     "snaps, expected_snap_call",
     [
@@ -110,13 +118,16 @@ def mock_inject_from_host(mocker):
 @pytest.mark.parametrize(
     "tag, expected_tag", [(None, "buildd-base-v0"), ("test-tag", "test-tag")]
 )
-def test_setup(  # pylint: disable=too-many-arguments
+def test_setup(  # pylint: disable=too-many-arguments, too-many-locals
     fake_process,
     fake_executor,
+    fake_filesystem,
     alias,
     hostname,
     environment,
     etc_environment_content,
+    no_cdn,
+    mock_load,
     mock_inject_from_host,
     mock_install_from_store,
     mocker,
@@ -136,6 +147,17 @@ def test_setup(  # pylint: disable=too-many-arguments
 
     if environment is None:
         environment = buildd.default_command_environment()
+
+    if no_cdn:
+        fake_filesystem.create_file(
+            "/etc/systemd/system/snapd.service.d/no-cdn.conf",
+            contents=dedent(
+                """\
+                [Service]
+                Environment=SNAPPY_STORE_NO_CDN=1
+                """
+            ),
+        )
 
     base_config = buildd.BuilddBase(
         alias=alias,
@@ -204,6 +226,10 @@ def test_setup(  # pylint: disable=too-many-arguments
     fake_process.register_subprocess(
         [*DEFAULT_FAKE_CMD, "systemctl", "start", "systemd-udevd"]
     )
+    if no_cdn:
+        fake_process.register_subprocess(
+            [*DEFAULT_FAKE_CMD, "mkdir", "-p", "/etc/systemd/system/snapd.service.d"]
+        )
     fake_process.register_subprocess(
         [*DEFAULT_FAKE_CMD, "apt-get", "install", "-y", "snapd"]
     )
@@ -216,10 +242,22 @@ def test_setup(  # pylint: disable=too-many-arguments
     fake_process.register_subprocess(
         [*DEFAULT_FAKE_CMD, "snap", "wait", "system", "seed.loaded"]
     )
+    fake_process.register_subprocess(
+        [*DEFAULT_FAKE_CMD, "snap", "set", "system", "proxy.http=http://foo.bar:8080"]
+    )
+    fake_process.register_subprocess(
+        [*DEFAULT_FAKE_CMD, "snap", "unset", "system", "proxy.http"]
+    )
+    fake_process.register_subprocess(
+        [*DEFAULT_FAKE_CMD, "snap", "set", "system", "proxy.https=http://foo.bar:8081"]
+    )
+    fake_process.register_subprocess(
+        [*DEFAULT_FAKE_CMD, "snap", "unset", "system", "proxy.https"]
+    )
 
     base_config.setup(executor=fake_executor)
 
-    assert fake_executor.records_of_push_file_io == [
+    expected_push_file_io = [
         dict(
             destination="/etc/apt/apt.conf.d/20auto-upgrades",
             content=dedent(
@@ -288,8 +326,18 @@ def test_setup(  # pylint: disable=too-many-arguments
             user="root",
         ),
     ]
+    expected_push_file = []
+    if no_cdn:
+        expected_push_file.append(
+            dict(
+                source=Path("/etc/systemd/system/snapd.service.d/no-cdn.conf"),
+                destination=Path("/etc/systemd/system/snapd.service.d/no-cdn.conf"),
+            )
+        )
+
+    assert fake_executor.records_of_push_file_io == expected_push_file_io
     assert fake_executor.records_of_pull_file == []
-    assert fake_executor.records_of_push_file == []
+    assert fake_executor.records_of_push_file == expected_push_file
     assert mock_install_from_store.mock_calls == expected_snap_call
 
 
@@ -719,12 +767,71 @@ def test_setup_resolved_restart_failure(
     )
 
 
+def test_setup_snapd_proxy(fake_executor, fake_process):
+    """Verify snapd proxy is set or unset."""
+    environment = dict(
+        http_proxy="http://foo.bar:8080",
+        https_proxy="http://foo.bar:8081",
+    )
+    base_config = buildd.BuilddBase(
+        alias=buildd.BuilddBaseAlias.FOCAL,
+        environment=environment,  # type: ignore
+    )
+    fake_process.keep_last_process(True)
+    fake_process.register([fake_process.any()])
+
+    base_config._setup_snapd_proxy(
+        executor=fake_executor,
+        deadline=None,
+    )
+    assert [
+        *DEFAULT_FAKE_CMD,
+        "snap",
+        "set",
+        "system",
+        "proxy.http=http://foo.bar:8080",
+    ] in fake_process.calls
+    assert [
+        *DEFAULT_FAKE_CMD,
+        "snap",
+        "set",
+        "system",
+        "proxy.https=http://foo.bar:8081",
+    ] in fake_process.calls
+
+
+@pytest.mark.parametrize("fail_index", list(range(0, 1)))
+def test_setup_snapd_proxy_failures(fake_process, fake_executor, fail_index):
+    base_config = buildd.BuilddBase(alias=buildd.BuilddBaseAlias.FOCAL)
+
+    return_codes = [0, 0]
+    return_codes[fail_index] = 1
+
+    fake_process.register_subprocess(
+        [*DEFAULT_FAKE_CMD, "snap", "unset", "system", "proxy.http"],
+        returncode=return_codes[0],
+    )
+    fake_process.register_subprocess(
+        [*DEFAULT_FAKE_CMD, "snap", "unset", "system", "proxy.https"],
+        returncode=return_codes[1],
+    )
+
+    with pytest.raises(errors.BaseConfigurationError) as exc_info:
+        base_config._setup_snapd_proxy(  # pylint: disable=protected-access
+            executor=fake_executor,
+            deadline=None,
+        )
+
+    assert exc_info.value == errors.BaseConfigurationError(
+        brief="Failed to set the snapd proxy.",
+        details=details_from_called_process_error(
+            exc_info.value.__cause__  # type: ignore
+        ),
+    )
+
+
 @pytest.mark.parametrize("fail_index", list(range(0, 7)))
-def test_setup_snapd_failures(
-    fake_process,
-    fake_executor,
-    fail_index,
-):
+def test_setup_snapd_failures(fake_process, fake_executor, fail_index):
     base_config = buildd.BuilddBase(alias=buildd.BuilddBaseAlias.FOCAL)
 
     return_codes = [0, 0, 0, 0, 0, 0, 0]
@@ -945,12 +1052,23 @@ def test_ensure_config_compatible_empty_config_returns_none(fake_executor):
     )
 
 
-def test_warmup_overall(fake_process, fake_executor, mock_load):
+@pytest.mark.parametrize(
+    "environment",
+    [
+        None,
+        dict(
+            https_proxy="http://foo.bar:8081",
+            http_proxy="http://foo.bar:8080",
+        ),
+    ],
+)
+def test_warmup_overall(environment, fake_process, fake_executor, mock_load):
     alias = buildd.BuilddBaseAlias.JAMMY
-    base_config = buildd.BuilddBase(
-        alias=alias,
-        environment=buildd.default_command_environment(),
-    )
+
+    if environment is None:
+        environment = buildd.default_command_environment()
+
+    base_config = buildd.BuilddBase(alias=alias, environment=environment)
 
     fake_process.register_subprocess(
         [*DEFAULT_FAKE_CMD, "cat", "/etc/os-release"],
@@ -968,6 +1086,18 @@ def test_warmup_overall(fake_process, fake_executor, mock_load):
     )
     fake_process.register_subprocess(
         [*DEFAULT_FAKE_CMD, "getent", "hosts", "snapcraft.io"]
+    )
+    fake_process.register_subprocess(
+        [*DEFAULT_FAKE_CMD, "snap", "set", "system", "proxy.http=http://foo.bar:8080"]
+    )
+    fake_process.register_subprocess(
+        [*DEFAULT_FAKE_CMD, "snap", "unset", "system", "proxy.http"]
+    )
+    fake_process.register_subprocess(
+        [*DEFAULT_FAKE_CMD, "snap", "set", "system", "proxy.https=http://foo.bar:8081"]
+    )
+    fake_process.register_subprocess(
+        [*DEFAULT_FAKE_CMD, "snap", "unset", "system", "proxy.https"]
     )
 
     base_config.warmup(executor=fake_executor)