+++ docs/changelog.rst
@@ -2,6 +2,18 @@
 Changelog
 *********
 
+1.5.0 (2022-09-23)
+------------------
+- Add mount method to Executor base class
+- LXDInstance's mount method signature has changed - The optional parameter `device_name` has been deprecated. It now matches MultipassInstance's signature of `mount(host_source, target)`
+- Signed snaps injected into a provider are asserted
+- Existing .snap files are not removed before overwriting with a new .snap file
+
+1.4.2 (2022-09-09)
+------------------
+- Set snapd http-proxy and https-proxy
+- Pass on snapd no-CDN configuration
+
 1.4.1 (2022-08-30)
 ------------------
 - Fix bug in BuilddBase where hostnames longer than 64 characters may